refactor: change from molecules to organisms the navbar
