+++ components/Organisms/Navbar/Navbar.tsx
@@ -1,7 +1,8 @@
 import Link from '@jpe-reader/components/Atoms/Link';
-import Button from '../../Atoms/Button';
+import Button from '@jpe-reader/components/Atoms/Button';
 import Title from '@jpe-reader/components/Atoms/Title';
-import LanguageSelection from '../LanguageSelection';
+import LanguageSelection from '@jpe-reader/components/Molecules/LanguageSelection';
+import SearchBar from '@jpe-reader/components/Molecules/SearchBar';
 import { useTheme } from '@jpe-reader/hooks/useTheme';
 import { useDictionary } from '@jpe-reader/hooks/useLanguage';
 import styles from './styles.module.css';
@@ -15,6 +16,7 @@ export default function Navbar() {
       <div className={styles.containerTitle}>
         <Title Component="h1">{navbar.title}</Title>
       </div>
+      <SearchBar />
       <nav className={styles.nav}>
         <ul className={styles.ul}>
           <li>
